extract Igc.Stats.RollingAverage
